requirements.txt
playwright
sentence-transformers>=2.2.2
faiss-cpu
nltk>=3.8
numpy>=1.24.0
torch>=2.0.0
langchain>=0.1.0
langgraph>=0.1.0
google-generativeai>=0.3.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
typing-extensions>=4.5.0
python-dotenv